+++ lib/rack/pretty_json.rb
@@ -4,31 +4,37 @@ module Rack
   class PrettyJSON
     def initialize(app, options = {})
       @app = app
-      @unless_header = (
-        options[:unless_header] || 'X-Rack-PrettyJSON-Skip'
+      @only_if_header = (
+        options[:only_if_header] || 'X-Rack-PrettyJSON'
       )
-      @unless_query = (
-        options[:unless_query] || 'compact'
+      @only_if_query = (
+        options[:only_if_query] || 'pretty'
       )
+      @warning = !!options[:warning]
     end
 
     def call(env)
-      p env
-      skip_client = env['HTTP_'+@unless_header.gsub(/-/,'_').upcase]
+      client = env['HTTP_'+@only_if_header.gsub(/-/,'_').upcase]
       
       code, head, body = @app.call(env)
 
-      skip_server = head.delete(@unless_query) || env['QUERY_STRING'].split("&").include?{|x| x.split("=")[0] == @unless_query}
+      server = head.delete(@only_if_header) || env['QUERY_STRING'].
+        split("&").find{|x| x.split("=")[0] == @only_if_query}
       
-      if !skip_client && !skip_server && head['Content-Type'] &&
-          head['Content-Type'].split(";")[0] =~ /[\+|\/]json$/i
-        payload = ""
-        body.each{|chunk| payload << chunk}
-        body = JSON.pretty_generate(
-          JSON.load(payload)
-        )
-        head['Content-Length'] = body.size.to_s
-        body = [body]
+      if head['Content-Type'] && head['Content-Type'].
+          split(";")[0] =~ /[\+|\/]json$/i
+        if (client || server) 
+          payload = ""
+          body.each{|chunk| payload << chunk}
+          body = JSON.pretty_generate(
+            JSON.load(payload)
+          )
+          head['Content-Length'] = body.size.to_s
+          body = [body]
+        elsif @warning
+          head['Warning'] ||= []
+          head['Warning'].push('Use `?pretty=yes` or add the HTTP header `X-Rack-PrettyJSON: yes` if you want pretty output.')
+        end
       end
 
       [code, head, body]